src/basic_classifiers/requirements.txt
certifi==2022.12.7
charset-normalizer==2.0.12
cycler==0.11.0
idna==3.4
imageio==2.34.0
joblib==1.1.1
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
pandas==1.0.1
Pillow==8.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==5.4.1
requests==2.27.0
scikit-learn==0.22.2.post1
scipy==1.5.4
seaborn==0.10.0
six==1.16.0
torch==1.8.0+cu101
torchvision==0.9.0+cu101
typing_extensions==4.1.1
urllib3==1.26.15
